apply review changes.
