+++ assets/src/edit-story/app/index.js
@@ -45,6 +45,7 @@ import { useLocalMedia, useMedia, MediaProvider } from './media';
 import { useStory, StoryProvider } from './story';
 import { useSnackbar, SnackbarProvider } from './snackbar';
 import Layout from './layout';
+import { Media3pApiProvider } from './media/media3p/api';
 
 function App({ config }) {
   const { storyId, isRTL } = config;
@@ -54,28 +55,30 @@ function App({ config }) {
         <ErrorBoundary>
           <ConfigProvider config={config}>
             <APIProvider>
-              <HistoryProvider size={50}>
-                <SnackbarProvider>
-                  <StoryProvider storyId={storyId}>
-                    <FontProvider>
-                      <MediaProvider>
-                        <AutoSaveHandler />
-                        <TransformProvider>
-                          <DropTargetsProvider>
-                            <GlobalStyle />
-                            <DevTools />
-                            <DefaultMoveableGlobalStyle />
-                            <CropMoveableGlobalStyle />
-                            <ModalGlobalStyle />
-                            <KeyboardOnlyOutlines />
-                            <Layout />
-                          </DropTargetsProvider>
-                        </TransformProvider>
-                      </MediaProvider>
-                    </FontProvider>
-                  </StoryProvider>
-                </SnackbarProvider>
-              </HistoryProvider>
+              <Media3pApiProvider>
+                <HistoryProvider size={50}>
+                  <SnackbarProvider>
+                    <StoryProvider storyId={storyId}>
+                      <FontProvider>
+                        <MediaProvider>
+                          <AutoSaveHandler />
+                          <TransformProvider>
+                            <DropTargetsProvider>
+                              <GlobalStyle />
+                              <DevTools />
+                              <DefaultMoveableGlobalStyle />
+                              <CropMoveableGlobalStyle />
+                              <ModalGlobalStyle />
+                              <KeyboardOnlyOutlines />
+                              <Layout />
+                            </DropTargetsProvider>
+                          </TransformProvider>
+                        </MediaProvider>
+                      </FontProvider>
+                    </StoryProvider>
+                  </SnackbarProvider>
+                </HistoryProvider>
+              </Media3pApiProvider>
             </APIProvider>
           </ConfigProvider>
         </ErrorBoundary>
