+++ assets/src/edit-story/components/library/libraryPanes.js
@@ -20,7 +20,6 @@
 import useLibrary from './useLibrary';
 import { getPane } from './panes';
 import { getTabId } from './panes/shared';
-import { Media3pApiProvider } from './panes/media/media3p/api';
 
 function LibraryPanes() {
   const { tab, tabs } = useLibrary((state) => ({
@@ -28,15 +27,9 @@ function LibraryPanes() {
     tabs: state.data.tabs,
   }));
   const panes = tabs.map(getPane);
-
-  // TODO(diegovar): Potentially move Media3pApiProvider down.
-  return (
-    <Media3pApiProvider>
-      {panes.map(({ id, Pane }) => (
-        <Pane key={id} isActive={id === tab} aria-labelledby={getTabId(id)} />
-      ))}
-    </Media3pApiProvider>
-  );
+  return panes.map(({ id, Pane }) => (
+    <Pane key={id} isActive={id === tab} aria-labelledby={getTabId(id)} />
+  ));
 }
 
 export default LibraryPanes;
